Added code for the Manufacturer name of vehicles
The difference between this one and the original is that it adds the "Brand" name in the .lua (example: Dodge), so you dont get "unknown" in game. This is creates a file that formats the vehicle list for copy and paste. HOWEVER you may want to change the vehicle class in the output.lua